Refactor retrieval of purchased course and summary details
This commit refactors the code in the `GetPurchasedCourse` and `GetPurchasedSummary` functions in the `Purchased.js` controller. It introduces the retrieval of all reviews for a course or summary, using the `Reviews` and `Review_Summary` models respectively. This change allows for a more comprehensive display of reviews in the API response.